+++ app/custom-orders/page.js
@@ -1,11 +1,11 @@
-import React from 'react'
-import OrderForm from '@/app/theme/fgstore.web/CustomOrder';
-import OrderFormApp from '@/app/theme/fgstore.mapp/home/CustomOrder';
+import React from "react";
+import OrderForm from "@/app/theme/fgstore.web/CustomOrder";
+import OrderFormApp from "@/app/theme/fgstore.mapp/home/CustomOrder";
+import { LocalSetup } from "../env";
 
 const page = () => {
-  return (
-    <OrderForm />
-  )
-}
+  if (LocalSetup === "fgstore.mapp") return <OrderFormApp />;
+  return <OrderForm />;
+};
 
-export default page
+export default page;