+++ app/env.js
@@ -1,3 +1,7 @@
 export const LocalSetup = "fgstore.mapp";
 // export const LocalSetup = "hoq.mapp";
 // export const LocalSetup = "fgstore.web";
+
+
+
+export const activeBrand = "EliorApp"